+++ .eslintrc.json
@@ -24,6 +24,7 @@
   "plugins": ["@typescript-eslint", "no-skip-tests", "no-only-tests"],
   "rules": {
     "import/prefer-default-export": "off",
+    "class-methods-use-this": "off",
     "max-classes-per-file": "off",
     "@typescript-eslint/explicit-module-boundary-types": "off",
     "@typescript-eslint/no-explicit-any": "off",
